requirements.txt
discord.py
aiogram
asyncio
